Refactor backend logic from SplitView and SessionView (#527)
* Pull out backend logic from split view

* Remove backend logic from SessionView

* Rename remove function
